+++ CHANGES.md
@@ -55,6 +55,9 @@ Use this section to keep track of changes in the works.
 ([lbluque](https://github.com/lbluque))
  
 ### Fixed
+* Disallow setting chemical potentials/fugacities with duplicate string/species
+  in dictionary. [\#114](https://github.com/CederGroupHub/smol/pull/114)
+  ([lbluque](https://github.com/lbluque))
 * Fixed loading `ClusterSubspace` with polynomial basis from dict.
 [\#112](https://github.com/CederGroupHub/smol/pull/112)
   ([lbluque](https://github.com/lbluque))